+++ cmd/gogoodwe/utils/jsonutils.go
@@ -1,5 +1,5 @@
 /*
-# Name: powerstationhelper - helper functions to get the Powerstation Data from the API
+# Name: jsonutils - helper functions to get the Powerstation Data from the API
 # Author: Aaron Saikovski - asaikovski@outlook.com
 */
 package utils
@@ -42,4 +42,16 @@ func GetDataJSON(PowerstationOutputData *types.InverterData) ([]byte, error) {
 	// Get the response and return any errors
 	resp, err := MarshalStructToJSON(&PowerstationOutputData)
 	return resp, err
+}
+
+// UnmarshalDataToStruct - Unmarshall http response to target struct
+func UnmarshalDataToStruct(respBody []byte, targetStruct interface{}) error {
+	resperr := json.Unmarshal(respBody, &targetStruct)
+	return resperr
+}
+
+// MarshalStructToJSON - Marshall the struct pointer to JSON
+func MarshalStructToJSON(targetStruct interface{}) ([]byte, error) {
+	jsonData, err := json.Marshal(&targetStruct)
+	return jsonData, err
 }
